rustc_abi
Module callconv

Modules

abi 🔒

Macros

reg_ctor 🔒

Structs

Heterogeneous

Error from the homogeneous_aggregate test function, indicating there are distinct leaf fields passed in different ways, or this is uninhabited.

Reg

Enums

HomogeneousAggregate

Return value from the homogeneous_aggregate test function.

RegKind