rustc_abi

Struct LayoutCalculator

pub struct LayoutCalculator<Cx> {
    pub cx: Cx,
}

Fields
cx: Cx
Implementations

impl<Cx: HasDataLayout> LayoutCalculator<Cx>

pub fn new(cx: Cx) -> Self

pub fn array_like<FieldIdx: Idx, VariantIdx: Idx, F>( &self, element: &LayoutData<FieldIdx, VariantIdx>, count_if_sized: Option<u64>, ) -> Result<LayoutData<FieldIdx, VariantIdx>, LayoutCalculatorError<F>>

pub fn simd_type<FieldIdx: Idx, VariantIdx: Idx, F: AsRef<LayoutData<FieldIdx, VariantIdx>> + Debug>( &self, element: F, count: u64, repr_packed: bool, ) -> Result<LayoutData<FieldIdx, VariantIdx>, LayoutCalculatorError<F>>

pub fn coroutine<'a, F: Deref<Target = &'a LayoutData<FieldIdx, VariantIdx>> + Debug + Copy, VariantIdx: Idx, FieldIdx: Idx, LocalIdx: Idx>( &self, local_layouts: &IndexSlice<LocalIdx, F>, prefix_layouts: IndexVec<FieldIdx, F>, variant_fields: &IndexSlice<VariantIdx, IndexVec<FieldIdx, LocalIdx>>, storage_conflicts: &BitMatrix<LocalIdx, LocalIdx>, tag_to_layout: impl Fn(Scalar) -> F, ) -> Result<LayoutData<FieldIdx, VariantIdx>, LayoutCalculatorError<F>>

Compute the layout for a coroutine.

This uses dedicated code instead of Self::layout_of_struct_or_enum, as coroutine fields may be shared between multiple variants (see the coroutine module for details).

pub fn univariant<'a, FieldIdx: Idx, VariantIdx: Idx, F: Deref<Target = &'a LayoutData<FieldIdx, VariantIdx>> + Debug + Copy>( &self, fields: &IndexSlice<FieldIdx, F>, repr: &ReprOptions, kind: StructKind, ) -> Result<LayoutData<FieldIdx, VariantIdx>, LayoutCalculatorError<F>>

pub fn layout_of_struct_or_enum<'a, FieldIdx: Idx, VariantIdx: Idx, F: Deref<Target = &'a LayoutData<FieldIdx, VariantIdx>> + Debug + Copy>( &self, repr: &ReprOptions, variants: &IndexSlice<VariantIdx, IndexVec<FieldIdx, F>>, is_enum: bool, is_special_no_niche: bool, scalar_valid_range: (Bound<u128>, Bound<u128>), discr_range_of_repr: impl Fn(i128, i128) -> (Integer, bool), discriminants: impl Iterator<Item = (VariantIdx, i128)>, always_sized: bool, ) -> Result<LayoutData<FieldIdx, VariantIdx>, LayoutCalculatorError<F>>

pub fn layout_of_union<'a, FieldIdx: Idx, VariantIdx: Idx, F: Deref<Target = &'a LayoutData<FieldIdx, VariantIdx>> + Debug + Copy>( &self, repr: &ReprOptions, variants: &IndexSlice<VariantIdx, IndexVec<FieldIdx, F>>, ) -> Result<LayoutData<FieldIdx, VariantIdx>, LayoutCalculatorError<F>>

fn layout_of_struct<'a, FieldIdx: Idx, VariantIdx: Idx, F: Deref<Target = &'a LayoutData<FieldIdx, VariantIdx>> + Debug + Copy>( &self, repr: &ReprOptions, variants: &IndexSlice<VariantIdx, IndexVec<FieldIdx, F>>, is_enum: bool, is_special_no_niche: bool, scalar_valid_range: (Bound<u128>, Bound<u128>), always_sized: bool, present_first: VariantIdx, ) -> Result<LayoutData<FieldIdx, VariantIdx>, LayoutCalculatorError<F>>

single-variant enums are just structs, if you think about it

fn layout_of_enum<'a, FieldIdx: Idx, VariantIdx: Idx, F: Deref<Target = &'a LayoutData<FieldIdx, VariantIdx>> + Debug + Copy>( &self, repr: &ReprOptions, variants: &IndexSlice<VariantIdx, IndexVec<FieldIdx, F>>, discr_range_of_repr: impl Fn(i128, i128) -> (Integer, bool), discriminants: impl Iterator<Item = (VariantIdx, i128)>, ) -> Result<LayoutData<FieldIdx, VariantIdx>, LayoutCalculatorError<F>>

fn univariant_biased<'a, FieldIdx: Idx, VariantIdx: Idx, F: Deref<Target = &'a LayoutData<FieldIdx, VariantIdx>> + Debug + Copy>( &self, fields: &IndexSlice<FieldIdx, F>, repr: &ReprOptions, kind: StructKind, niche_bias: NicheBias, ) -> Result<LayoutData<FieldIdx, VariantIdx>, LayoutCalculatorError<F>>

fn format_field_niches<'a, FieldIdx: Idx, VariantIdx: Idx, F: Deref<Target = &'a LayoutData<FieldIdx, VariantIdx>> + Debug>( &self, layout: &LayoutData<FieldIdx, VariantIdx>, fields: &IndexSlice<FieldIdx, F>, ) -> String