Struct rustc_abi::ReprOptions

pub struct ReprOptions {
    pub int: Option<IntegerType>,
    pub align: Option<Align>,
    pub pack: Option<Align>,
    pub flags: ReprFlags,
    pub field_shuffle_seed: Hash64,
}

Represents the repr options provided by the user,

Fields
int: Option<IntegerType>align: Option<Align>pack: Option<Align>flags: ReprFlagsfield_shuffle_seed: Hash64

The seed to be used for randomizing a type’s layout

Note: This could technically be a Hash128 which would be the “most accurate” hash as it’d encompass the item and crate hash without loss, but it does pay the price of being larger. Everything’s a tradeoff, a 64-bit seed should be sufficient for our purposes (primarily -Z randomize-layout)

Implementations

impl ReprOptions

pub fn simd(&self) -> bool

pub fn c(&self) -> bool

pub fn packed(&self) -> bool

pub fn transparent(&self) -> bool

pub fn linear(&self) -> bool

pub fn discr_type(&self) -> IntegerType

Returns the discriminant type, given these repr options. This must only be called on enums!

pub fn inhibit_enum_layout_opt(&self) -> bool

Returns true if this #[repr()] should inhabit “smart enum layout” optimizations, such as representing Foo<&T> as a single pointer.

pub fn inhibit_struct_field_reordering_opt(&self) -> bool

Returns true if this #[repr()] should inhibit struct field reordering optimizations, such as with repr(C), repr(packed(1)), or repr(<int>).

pub fn can_randomize_type_layout(&self) -> bool

Returns true if this type is valid for reordering and -Z randomize-layout was enabled for its declaration crate

pub fn inhibit_union_abi_opt(&self) -> bool

Returns true if this #[repr()] should inhibit union ABI optimisations.