rustc_abi::layout
Module ty

Structs

FieldIdx

The source-order index of a field in a variant.

Layout

TyAndLayout

The layout of a type, alongside the type itself. Provides various type traversal APIs (e.g., recursing into fields).

VariantIdx

The source-order index of a variant in a type.

Constants

FIRST_VARIANT

Equivalent to VariantIdx(0).

Traits

TyAbiInterface

Trait that needs to be implemented by the higher-level type representation (e.g. rustc_middle::ty::Ty), to provide rustc_target::abi functionality.